Even More Joust Polish

+++ firmware/user/display/font.h
@@ -29,7 +29,7 @@ extern const sprite_t font_IbmVga8[] RODATA_ATTR;
 #define FONT_HEIGHT_TOMTHUMB 5
 extern const sprite_t font_TomThumb[] RODATA_ATTR;
 
-uint8_t plotChar(uint8_t x, uint8_t y, char character, const sprite_t* table, color col);
-uint8_t plotText(uint8_t x, uint8_t y, char* text, fonts font, color col);
+int16_t plotChar(int16_t x, int16_t y, char character, const sprite_t* table, color col);
+int16_t plotText(int16_t x, int16_t y, char* text, fonts font, color col);
 
 #endif /* SRC_FONT_H_ */
